requirements-dev.txt
coppyr[dev]
sanic-testing
